tests/requirements.txt
djangorestframework==3.15.1
django-extensions==3.2.1
Django==4.1.7
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
